WIP: Delete _containers.scss. Begin streamlining SCSS styling for content spacing by updating _section-components.scss and _type.scss
The scss approach to content spacing is too complicated right now. Honestly it's a mess and there's a lack of design consistency. I have been relying on a gap value for content spacing in <section>, but each section has h2 and h3 headers, which require different margin-bottoms, and I didn't want to add a css class to each, along with all the other components—I'd rather change my career.

Progress so far:
- Removed gap value from .content, .section.container
- Reset margin-bottom to all section heading and paragraph elements to override defaults
- Updated all heading and paragraph elements to have margin top and bottom
In _type.scss, I left a margin-bottom value of 1.5rem to <h2> to make space for