remove ./dist
It's generally a good practice to include the dist (distribution) directory in your .gitignore file.

The dist directory is usually generated by your build tools and contains the compiled version of your code. This directory can be easily regenerated from your source code, so it's not necessary to include it in your version control system.

Including it could lead to unnecessary merge conflicts and would bloat the size of your repository. Instead, you should only include the source code and the instructions for generating the dist directory (like a build script in your package.json file).